More refactoring: moved tabs into http_request_inspector_views; refactoring network_screen's build method
